+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Encoders:
 - Half adder and full adder gate constraint encoder
 - Bitvector AND, OR, XOR gate constraint encoder
 - Ripple-carry bitvector adder gate constraint encoder
+- Ripple-carry bitvector subtractor gate constraint encoder (for 2's complement signed integers)
 - Parallel bitvector multiplier gate constraint encoder
 - Signed (2's complement) and unsigned bitvector comparison gate constraint encoder